# Docs/openapi-nfe-entrada-v1.yaml
...
openapi: 3.0.3
info:
  title: EXEQ Hub NF-e Entrada
  version: 1.0.0-draft
...
paths:
  /nfe/entrada/:
    get:
      tags: [nfe_entrada]
      summary: Lista documentos recebidos via distNSU
      description: |
        Filtros por emitente/chave, manifestação, status XML, provider e período.
        Resposta paginada (HubPageNumberPagination) com KPIs agregados no mesmo payload.
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Busca livre (emitente, CNPJ, chave, número)
        - in: query
          name: manifest_status
          schema:
            type: string
            enum: [all, none, ciencia, confirmada, desconhecida, nao_realizada]
        - in: query
          name: xml_status
          schema:
            type: string
            enum: [all, pending, available, error]
        - in: query
          name: provider_id
          schema: { type: string, format: uuid }
        - in: query
          name: days
          schema:
            type: integer
            description: "Janela em issue_date; 0 = sem limite; default 30 se apply_default_period"
        - in: query
          name: date_from
          schema: { type: string, format: date }
        - in: query
          name: date_to
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: Página + KPIs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NfeEntradaListPage" }
        "403":
          description: NF-e de entrada desabilitada (global ou tenant)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NfeEntradaError" }
...
components:
...
  schemas:
    NfeEntradaError:
      type: object
      properties:
        detail: { type: string }
        code:
          type: string
          description: "nfe_entrada_disabled | nfe_entrada_manifestation | …"
    NfeEntradaKpis:
      type: object
      properties:
        total: { type: integer }
        xml_pending: { type: integer }
        manifest_pending: { type: integer }
        xml_available: { type: integer }
    NfeEntradaListPage:
      type: object
      description: Paginação DRF + kpis no root
      properties:
        count: { type: integer }
        next: { type: string, format: uri, nullable: true }
        previous: { type: string, format: uri, nullable: true }
        results:
          type: array
          items: { $ref: "#/components/schemas/NfeEntradaDocument" }
        kpis: { $ref: "#/components/schemas/NfeEntradaKpis" }
    NfeEntradaDocument:
      type: object
      properties:
        id: { type: string, format: uuid }
        provider: { type: string, format: uuid }
        provider_name: { type: string }
        nsu: { type: string }
        schema_type:
          type: string
          enum: [resNFe, procNFe, resEvento, procEventoNFe, other]
        access_key: { type: string, maxLength: 44 }
        issuer_cnpj: { type: string }
        issuer_name: { type: string }
        recipient_cnpj: { type: string }
        number: { type: integer, nullable: true }
        series: { type: integer, nullable: true }
        issue_date: { type: string, format: date, nullable: true }
        total_cents: { type: integer, nullable: true }
        nfe_status: { type: string }
        xml_status:
          type: string
          enum: [pending, available, error]
        manifest_status:
          type: string
          enum: [none, ciencia, confirmada, desconhecida, nao_realizada]
        has_xml: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...